+++ lib/service/Health.dart
@@ -50,6 +50,8 @@ class Health with Service implements NotificationsListener {
   static const String notifyHistoryUpdated          = "edu.illinois.rokwire.health.history.updated";
   static const String notifyUserUpdated             = "edu.illinois.rokwire.health.user.updated";
   static const String notifyUserPrivateKeyUpdated   = "edu.illinois.rokwire.health.user.private_key.updated";
+  static const String notifyFamilyMembersAvailable  = "edu.illinois.rokwire.health.family.members.available";
+  static const String notifyPendingFamilyMember     = "edu.illinois.rokwire.health.family.member.pending";
   
   static const String notifyProcessingFinished      = "edu.illinois.rokwire.health.processing.finished";
   
@@ -113,7 +115,9 @@ class Health with Service implements NotificationsListener {
 
   @override
   void initServiceUI() {
-    _process();
+    _process().then((_) {
+      checkPendingFamilyMembers();
+    });
   }
 
   @override
@@ -163,6 +167,7 @@ class Health with Service implements NotificationsListener {
         if (Config().refreshTimeout < pausedDuration.inSeconds) {
           _refreshUser().then((_) {
             _process();
+            checkPendingFamilyMembers();
           });
         }
       }
@@ -1624,6 +1629,54 @@ class Health with Service implements NotificationsListener {
       return true;
     }
     return false;
+  }
+
+
+  // Membership application
+
+  Future<List<HealthFamilyMember>> loadFamilyMembers() async {
+    String url = (Config().healthUrl != null) ? "${Config().healthUrl}/covid19/join-external-approvements" : null;
+    Response response = (url != null) ? await Network().get(url, auth: NetworkAuth.User) : null;
+    String responseBody = (response?.statusCode == 200) ? response.body : null;
+    List<dynamic> responseJson = (responseBody != null) ? AppJson.decodeList(responseBody) : null;
+    /*responseJson = [
+      { "id": "1234", "first_name": "Petyo", "last_name": "Stoyanov", "date_created": "2021-02-19T10:27:43.679Z", "group_name": "U of Illinois employee family member", "external_approver_id": "68572", "external_approver_last_name": "Varbanov", "status":"pending" },
+      { "id": "5678", "first_name": "Mladen", "last_name": "Dryankov", "date_created": "2021-02-19T10:27:43.679Z", "group_name": "U of Illinois employee family member", "external_approver_id": "68572", "external_approver_last_name": "Varbanov", "status":"accepted" },
+      { "id": "9101", "first_name": "Todor", "last_name": "Bachvarov", "date_created": "2021-02-19T10:27:43.679Z", "group_name": "U of Illinois employee family member", "external_approver_id": "68572", "external_approver_last_name": "Varbanov", "status":"rejected" },
+    ];*/
+
+    return  (responseJson != null) ? HealthFamilyMember.listFromJson(responseJson) : null;
+  }
+
+  Future<void> checkPendingFamilyMembers() async {
+    if (Storage().onBoardingPassed && _isLoggedIn) {
+      List<HealthFamilyMember> members = await loadFamilyMembers();
+      if (members != null) {
+        NotificationService().notify(notifyFamilyMembersAvailable, members); 
+
+        HealthFamilyMember pendingMember = HealthFamilyMember.pendingMemberFromList(members);
+        if (pendingMember != null) {
+          NotificationService().notify(notifyPendingFamilyMember, pendingMember); 
+        }
+      }
+    }
+  }
+
+  // Return:
+  //  - true, on succeess
+  //  - false or error string on fail
+  Future<dynamic> applyFamilyMemberStatus(HealthFamilyMember member, String status) async {
+    String url = (Config().healthUrl != null) ? "${Config().healthUrl}/covid19/join-external-approvements/${member?.id}" : null;
+    String post = AppJson.encode({'status': status });
+    Response response = (url != null) ? await Network().put(url, body: post, auth: NetworkAuth.User) : null;
+
+    if (response?.statusCode == 200) {
+      return true;
+    }
+    else {
+      String responseBody = response?.body;
+      return (AppString.isStringNotEmpty(responseBody)) ? responseBody : false;
+    }
   }
 }