+++ lib/ui/RootPanel.dart
@@ -26,6 +26,7 @@ import 'package:illinois/service/Service.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Localization.dart';
 import 'package:illinois/service/NotificationService.dart';
+import 'package:illinois/ui/settings/SettingsPendingFamilyMemberPanel.dart';
 import 'package:illinois/ui/health/Covid19HistoryPanel.dart';
 import 'package:illinois/ui/health/Covid19InfoCenterPanel.dart';
 import 'package:illinois/ui/health/Covid19StatusPanel.dart';
@@ -50,6 +51,8 @@ class _RootPanelState extends State<RootPanel> with SingleTickerProviderStateMix
 
   static const String HEALTH_STATUS_URI = 'edu.illinois.covid://covid.illinois.edu/health/status';
 
+  bool _presentingMemberApplication;
+
   @override
   void initState() {
     super.initState();
@@ -60,6 +63,7 @@ class _RootPanelState extends State<RootPanel> with SingleTickerProviderStateMix
       Organizations.notifyOrganizationChanged,
       Organizations.notifyEnvironmentChanged,
       Health.notifyStatusUpdated,
+      Health.notifyPendingFamilyMember,
       DeepLink.notifyUri,
     ]);
 
@@ -90,6 +94,9 @@ class _RootPanelState extends State<RootPanel> with SingleTickerProviderStateMix
     }
     else if (name == Health.notifyStatusUpdated) {
       _presentHealthStatusUpdate(param);
+    }
+    else if (name == Health.notifyPendingFamilyMember) {
+      _presentPedningFamilyMember(param);
     }
     else if (name == FirebaseMessaging.notifyCovid19Notification) {
       _onFirebaseCovid19Notification(param);
@@ -224,6 +231,18 @@ class _RootPanelState extends State<RootPanel> with SingleTickerProviderStateMix
     Navigator.push(context, CupertinoPageRoute(builder: (context) => Covid19StatusUpdatePanel(status: params['status'], previousHealthStatus: params['lastHealthStatus'],)));
   }
 
+  void _presentPedningFamilyMember(dynamic param) {
+    if (_presentingMemberApplication != true) {
+      _presentingMemberApplication = true;
+      Navigator.push(context, PageRouteBuilder( opaque: false, pageBuilder: (context, _, __) => SettingsPendingFamilyMemberPanel(pendingMember: param))).then((result) {
+        _presentingMemberApplication = false;
+        if (result == true) {
+          Health().checkPendingFamilyMembers();
+        }
+      });
+    }
+  }
+  
   void _onDeeplinkUri(Uri uri) {
     if (uri != null) {
       Uri healthStatusUri;